chore(core): rename createSliceServices tool
