+++ plugins-structure/packages/core/src/toolkit/createSliceServices.test.js
@@ -1,4 +1,4 @@
-import { createSliceServices } from "./createSliceService";
+import { createSliceServices } from "./createSliceServices";
 
 const validServices = {
   foo: {
