+++ plugins-structure/packages/core/src/toolkit/index.js
@@ -1,1 +1,1 @@
-export * from "./createSliceService";
+export * from "./createSliceServices";
